Refactor badge and form components: remove unused badgeVariants and form field context, simplifying the codebase. Update sidebar to utilize the new useSidebar hook for improved context management.
